+++ server/odc-service/src/main/java/com/oceanbase/odc/service/quartz/QuartzJobService.java
@@ -15,47 +15,26 @@
  */
 package com.oceanbase.odc.service.quartz;
 
-import java.util.List;
+import org.quartz.Scheduler;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
 
-import org.quartz.JobDataMap;
-import org.quartz.JobKey;
-import org.quartz.SchedulerException;
-import org.quartz.Trigger;
-import org.quartz.TriggerKey;
-import org.quartz.UnableToInterruptJobException;
-
-import com.oceanbase.odc.service.schedule.model.ChangeQuartJobParam;
-import com.oceanbase.odc.service.schedule.model.CreateQuartzJobParam;
+import com.oceanbase.odc.core.authority.util.SkipAuthorize;
 
 /**
  * @author jingtian
  * @date 2024/12/25
  */
-public interface QuartzJobService {
-    void createJob(CreateQuartzJobParam req) throws SchedulerException;
+@Service("quartzJobService")
+@SkipAuthorize("odc internal usage")
+public class QuartzJobService extends AbstractQuartzJobService {
+    @Autowired(required = false)
+    @Qualifier(value = ("commonScheduler"))
+    private Scheduler commonScheduler;
 
-    void createJob(CreateQuartzJobParam req, JobDataMap triggerDataMap) throws SchedulerException;
-
-    void changeJob(ChangeQuartJobParam req);
-
-    void pauseJob(JobKey jobKey) throws SchedulerException;
-
-    void resumeJob(JobKey jobKey) throws SchedulerException;
-
-    void deleteJob(JobKey jobKey) throws SchedulerException;
-
-    boolean checkExists(JobKey jobKey) throws SchedulerException;
-
-    Trigger getTrigger(TriggerKey key) throws SchedulerException;
-
-    void rescheduleJob(TriggerKey triggerKey, Trigger newTrigger) throws SchedulerException;
-
-    void triggerJob(JobKey key) throws SchedulerException;
-
-    void interruptJob(JobKey key) throws UnableToInterruptJobException;
-
-    void triggerJob(JobKey key, JobDataMap triggerDataMap) throws SchedulerException;
-
-    List<? extends Trigger> getJobTriggers(JobKey jobKey) throws SchedulerException;
-
+    @Override
+    protected Scheduler getScheduler() {
+        return commonScheduler;
+    }
 }
